fix tendermint block to SequencerBlockData conversion (#424)
## Summary
fix tendermint block to SequencerBlockData conversion by fixing
`data_hash` calculation, txs need to be hashed before `data_hash` is
calculated

## Background 
our data_hash calculation was wrong, which caused the type conversion to
always fail. also moved a test which was previously in relayer which
tested this to astria-sequencer-types
