+++ crates/astria-sequencer-relayer/src/lib.rs
@@ -4,7 +4,6 @@ pub mod data_availability;
 pub(crate) mod macros;
 pub mod relayer;
 pub mod sequencer_relayer;
-pub mod transaction;
 pub mod validator;
 
 pub use sequencer_relayer::SequencerRelayer;
