+++ crates/astria-sequencer-relayer/tests/blackbox/helper.rs
@@ -310,6 +310,7 @@ impl StateServer for StateCelestiaImpl {
 
 fn create_block_response(validator: &Validator, height: u32) -> endpoint::block::Response {
     use proto::Message as _;
+    use sha2::Digest as _;
     use tendermint::{
         block,
         chain,
@@ -346,7 +347,7 @@ fn create_block_response(validator: &Validator, height: u32) -> endpoint::block:
         signed_tx_bytes,
     ];
     let data_hash = Some(Hash::Sha256(simple_hash_from_byte_vectors::<sha2::Sha256>(
-        &data,
+        &data.iter().map(sha2::Sha256::digest).collect::<Vec<_>>(),
     )));
 
     let (last_commit_hash, last_commit) = sequencer_types::test_utils::make_test_commit_and_hash();
